requirements.txt
osiris-sdk
apache-beam
pandas
azure-storage-file-datalake
